+++ src/components/Spreadsheet/data/types.d.ts
@@ -5,7 +5,8 @@ export type DownloadOptions = {
   id: string
   value: ExportValue
   file: File
-  separation: string
+  separation: Separation
 }
 export type File = 'csv' | 'plain/text'
 export type ExportValue = 'value' | 'expression'
+export type Separation = '\t' | ','
