+++ src/components/Spreadsheet/logic/useClipboard.ts
@@ -1,10 +1,10 @@
 import { parentTag } from '@/components/Spreadsheet/data/constants'
 import { HTMLCell } from '@/components/Spreadsheet/data/types'
+import { updateSelectedCellsValues } from '@/components/Spreadsheet/utils/cell'
 import {
   addTextToCellValues,
   formatCellValuesToText,
-  updateSelectedCellsValues,
-} from '@/components/Spreadsheet/logic/cellUtils'
+} from '@/components/Spreadsheet/utils/format'
 import { useEffect } from 'react'
 
 export const useClipboard = (selectedElements: NodeListOf<HTMLCell>) => {
@@ -13,18 +13,12 @@ export const useClipboard = (selectedElements: NodeListOf<HTMLCell>) => {
       if (!selectedElements) return
 
       // TO_DO: should be available as copy option (contextual menu)
-      const copyPlainText = false
-      const result = formatCellValuesToText(selectedElements, copyPlainText)
+      const result = formatCellValuesToText({ elements: selectedElements })
 
       event.clipboardData?.setData('text/plain', result)
       event.preventDefault()
     }
 
-    document.addEventListener('copy', copy)
-    return () => document.removeEventListener('copy', copy)
-  }, [selectedElements])
-
-  useEffect(() => {
     const paste = (event: ClipboardEvent) => {
       const clipboardData = event.clipboardData?.getData('text/plain')
       if (!clipboardData) return
@@ -37,17 +31,11 @@ export const useClipboard = (selectedElements: NodeListOf<HTMLCell>) => {
       element.focus()
     }
 
-    document.addEventListener('paste', paste)
-    return () => document.removeEventListener('paste', paste)
-  }, [selectedElements])
-
-  useEffect(() => {
     const cut = (event: ClipboardEvent) => {
       if (!selectedElements) return
 
-      // TO_DO: should be available as cut option (contextual menu)
-      const copyPlainText = false
-      const result = formatCellValuesToText(selectedElements, copyPlainText)
+      // TO_DO: should be available as cut option (contextual menu / config)
+      const result = formatCellValuesToText({ elements: selectedElements })
       const element = document.activeElement as HTMLCell
 
       updateSelectedCellsValues('', element, selectedElements)
@@ -56,7 +44,13 @@ export const useClipboard = (selectedElements: NodeListOf<HTMLCell>) => {
       event.preventDefault()
     }
 
+    document.addEventListener('copy', copy)
+    document.addEventListener('paste', paste)
     document.addEventListener('cut', cut)
-    return () => document.removeEventListener('cut', cut)
+    return () => {
+      document.removeEventListener('copy', copy)
+      document.removeEventListener('paste', paste)
+      document.removeEventListener('cut', cut)
+    }
   }, [selectedElements])
 }
