+++ src/components/Spreadsheet/logic/useMouse.ts
@@ -1,7 +1,7 @@
 import { parentTag } from '@/components/Spreadsheet/data/constants'
 import { HTMLCell } from '@/components/Spreadsheet/data/types'
 
-import { useCallback, useEffect, useRef } from 'react'
+import { useEffect, useRef } from 'react'
 
 export function useMouse(
   selectedElements: NodeListOf<HTMLCell>,
@@ -9,10 +9,10 @@ export function useMouse(
   removeSelection: () => void
 ) {
   const firstElement = useRef<HTMLCell | null>(null)
-  const isDrag = useRef<boolean>(false)
+  const isMovingAndDown = useRef<boolean>(false)
 
-  const mouseDown = useCallback(
-    (event: MouseEvent) => {
+  useEffect(() => {
+    const mouseDown = (event: MouseEvent) => {
       event.stopPropagation()
 
       const target = (event.target as HTMLElement).parentElement as HTMLCell
@@ -25,44 +25,35 @@ export function useMouse(
       if (isTargetSelected) return
 
       firstElement.current = target
-    },
-    [selectedElements]
-  )
+    }
 
-  const mouseDrag = useCallback(
-    (event: MouseEvent) => {
+    const mouseMove = (event: MouseEvent) => {
       event.preventDefault()
       event.stopPropagation()
 
       const target = (event.target as HTMLElement).parentElement as HTMLCell
       if (!firstElement.current || target.tagName !== parentTag) return
-      isDrag.current = true
+      isMovingAndDown.current = true
 
       addSelectionArea(firstElement.current, target)
-    },
-    [addSelectionArea]
-  )
+    }
 
-  const mouseUp = useCallback(
-    (event: MouseEvent) => {
+    const mouseUp = (event: MouseEvent) => {
       event.preventDefault()
       event.stopPropagation()
 
-      if (!isDrag.current) removeSelection()
+      if (!isMovingAndDown.current) removeSelection()
       firstElement.current = null
-      isDrag.current = false
-    },
-    [removeSelection]
-  )
+      isMovingAndDown.current = false
+    }
 
-  useEffect(() => {
     document.addEventListener('mousedown', mouseDown)
-    document.addEventListener('mousemove', mouseDrag)
+    document.addEventListener('mousemove', mouseMove)
     document.addEventListener('mouseup', mouseUp)
     return () => {
       document.removeEventListener('mousedown', mouseDown)
-      document.removeEventListener('mousemove', mouseDrag)
+      document.removeEventListener('mousemove', mouseMove)
       document.removeEventListener('mouseup', mouseUp)
     }
-  }, [mouseDrag, mouseDown, mouseUp])
+  }, [selectedElements, addSelectionArea, removeSelection])
 }
