+++ src/components/Spreadsheet/logic/useSelection.ts
@@ -5,7 +5,8 @@ import {
   getCell,
   getCurrentCellCoordinates,
   getOutput,
-} from '@/components/Spreadsheet/logic/cellUtils'
+} from '@/components/Spreadsheet/utils/cell'
+
 import { MouseEvent, useCallback, useEffect, useState } from 'react'
 
 export function useSelection() {
@@ -33,24 +34,25 @@ export function useSelection() {
     firstCell?.focus()
   }, [selectedElements])
 
-  const addDraggable = (elements: HTMLElement[]) => {
-    elements.forEach((el) => {
-      el?.classList.add('drag')
-      el?.setAttribute('draggable', 'true')
-    })
-  }
-  const removeDraggable = (elements: HTMLElement[]) => {
-    elements.forEach((el) => {
-      el?.classList.remove('drag')
-      el?.setAttribute('draggable', 'false')
-    })
-  }
-
   useEffect(() => {
     if (!selectedElements) return
     const draggedElements = Array.from(selectedElements).map((el) =>
       getOutput(el)
     )
+
+    const addDraggable = (elements: HTMLElement[]) => {
+      elements.forEach((el) => {
+        el?.classList.add('drag')
+        el?.setAttribute('draggable', 'true')
+      })
+    }
+
+    const removeDraggable = (elements: HTMLElement[]) => {
+      elements.forEach((el) => {
+        el?.classList.remove('drag')
+        el?.setAttribute('draggable', 'false')
+      })
+    }
 
     addDraggable(draggedElements)
     return () => removeDraggable(draggedElements)
