+++ src/context/matrix/MatrixProvider.tsx
@@ -5,7 +5,7 @@ import { MouseEvent, ReactNode, useCallback, useState } from 'react'
 import ComputedMatrix from '@/context/matrix/ComputedMatrix'
 import { v4 as uuidv4 } from 'uuid'
 import { toast } from '@/components/ui/toast'
-import { downloadTable } from '@/components/Spreadsheet/logic/cellUtils'
+import { downloadTable } from '@/components/Spreadsheet/utils/cellUtils'
 import { DownloadOptions } from '@/components/Spreadsheet/data/types'
 
 type Params = {
